Add move ReceiptPool out of query_engine
