+++ graph-gateway/src/query_engine.rs
@@ -7,32 +7,29 @@ use crate::{
     manifest_client::SubgraphInfo,
     metrics::*,
     price_automation::*,
+    receipts::*,
     unattestable_errors::UNATTESTABLE_ERROR_MESSAGE_FRAGMENTS,
 };
 use futures::future::join_all;
 use indexer_selection::{
-    self,
-    actor::IndexerErrorObservation,
-    receipts::{BorrowFail, QueryStatus as ReceiptStatus, ReceiptPool},
-    Context, FreshnessRequirements, IndexerError, IndexerPreferences, IndexerScore, ScoringSample,
-    Selection, SelectionError, UnresolvedBlock,
+    self, actor::IndexerErrorObservation, Context, FreshnessRequirements, IndexerError,
+    IndexerPreferences, IndexerScore, ScoringSample, Selection, SelectionError, UnresolvedBlock,
 };
 use indexer_selection::{actor::Update, Indexing, UtilityConfig};
 use lazy_static::lazy_static;
 use prelude::{buffer_queue::QueueWriter, double_buffer::DoubleBufferReader, graphql::Response, *};
 use primitive_types::U256;
-use secp256k1::SecretKey;
 use serde::Deserialize;
 use serde_json::value::RawValue;
 use std::{
-    collections::{hash_map::Entry, BTreeSet, HashMap},
+    collections::{BTreeSet, HashMap},
     rc::Rc,
     sync::{
         atomic::{AtomicUsize, Ordering as MemoryOrdering},
         Arc,
     },
 };
-use tokio::sync::{Mutex, RwLock};
+use tokio::sync::Mutex;
 use uuid::Uuid;
 
 #[derive(Debug)]
@@ -200,77 +197,6 @@ impl From<UnresolvedBlock> for QueryEngineError {
 pub struct Config {
     pub indexer_selection_retry_limit: usize,
     pub budget_factors: QueryBudgetFactors,
-}
-
-#[derive(Clone, Default)]
-pub struct ReceiptPools {
-    pools: Arc<RwLock<HashMap<Indexing, Arc<Mutex<ReceiptPool>>>>>,
-}
-
-impl ReceiptPools {
-    async fn get(&self, indexing: &Indexing) -> Arc<Mutex<ReceiptPool>> {
-        if let Some(pool) = self.pools.read().await.get(indexing) {
-            return pool.clone();
-        }
-        let mut pools = self.pools.write().await;
-        match pools.entry(indexing.clone()) {
-            Entry::Occupied(entry) => entry.get().clone(),
-            Entry::Vacant(entry) => {
-                let pool = Arc::new(Mutex::default());
-                entry.insert(pool.clone());
-                pool
-            }
-        }
-    }
-
-    pub async fn commit(&self, indexing: &Indexing, fee: GRT) -> Result<Vec<u8>, BorrowFail> {
-        let pool = self
-            .pools
-            .read()
-            .await
-            .get(indexing)
-            .cloned()
-            .ok_or(BorrowFail::NoAllocation)?;
-        let mut pool = pool.lock().await;
-        pool.commit(fee.shift::<0>().as_u256())
-    }
-
-    pub async fn release(&self, indexing: &Indexing, receipt: &[u8], status: ReceiptStatus) {
-        if receipt.len() != 164 {
-            panic!("Unrecognized receipt format");
-        }
-        let pool = self.pools.read().await;
-        let mut pool = match pool.get(indexing) {
-            Some(pool) => pool.lock().await,
-            None => return,
-        };
-        pool.release(receipt, status);
-    }
-
-    pub async fn update_receipt_pool(
-        &self,
-        signer: &SecretKey,
-        indexing: &Indexing,
-        new_allocations: &HashMap<Address, GRT>,
-    ) {
-        let pool = self.get(indexing).await;
-        let mut pool = pool.lock().await;
-        // Remove allocations not present in new_allocations
-        for old_allocation in pool.addresses() {
-            if new_allocations
-                .iter()
-                .all(|(id, _)| &old_allocation != id.as_ref())
-            {
-                pool.remove_allocation(&old_allocation);
-            }
-        }
-        // Add new_allocations not present in allocations
-        for (id, _) in new_allocations {
-            if !pool.contains_allocation(&id) {
-                pool.add_allocation(signer.clone(), id.0);
-            }
-        }
-    }
 }
 
 pub struct QueryEngine<I, K, F>
